Rename SpeakerAdapter -> SearchAdapter
